add mailer tests
