+++ src/controllers/notifications.controller.js
@@ -1,5 +1,4 @@
 const notificationsService = require('../services/notifications.service');
-const emailService = require('../services/email.service');
 
 exports.subscribe = async (req, res) => {
   try {
@@ -51,18 +50,3 @@ exports.send = async (req, res) => {
   }
 };
 
-exports.adminSend = async (req, res) => {
-  try {
-    const { subject, body, html } = req.body || {};
-    const { stats, marketplace } = req.query || {};
-    const emails = await emailService.getSubscribedEmails({
-      stats: stats === 'true',
-      marketplace: marketplace === 'true'
-    });
-    const result = await emailService.sendBulkEmail(emails, subject, body, html);
-    res.json(result);
-  } catch (err) {
-    console.error('Erreur envoi notifications admin:', err);
-    res.status(500).json({ error: "Erreur lors de l'envoi des notifications" });
-  }
-};
