+++ tests/mailer.test.js
@@ -1,0 +1,70 @@
+jest.mock('nodemailer', () => ({
+  createTransport: jest.fn(),
+}));
+let nodemailer;
+
+describe('mailer', () => {
+  beforeEach(() => {
+    jest.resetModules();
+    jest.clearAllMocks();
+    nodemailer = require('nodemailer');
+  });
+
+  test('creates Gmail OAuth2 transporter', () => {
+    process.env.EMAIL_USER = 'user@gmail.com';
+    process.env.GOOGLE_CLIENT_ID = 'clientId';
+    process.env.GOOGLE_CLIENT_SECRET = 'clientSecret';
+    process.env.GOOGLE_REFRESH_TOKEN = 'refresh';
+
+    require('../src/utils/mailer');
+
+    expect(nodemailer.createTransport).toHaveBeenCalledWith({
+      service: 'gmail',
+      auth: {
+        type: 'OAuth2',
+        user: process.env.EMAIL_USER,
+        clientId: process.env.GOOGLE_CLIENT_ID,
+        clientSecret: process.env.GOOGLE_CLIENT_SECRET,
+        refreshToken: process.env.GOOGLE_REFRESH_TOKEN,
+      },
+    });
+  });
+
+  test('sendMail uses EMAIL_USER when EMAIL_FROM missing', async () => {
+    process.env.EMAIL_USER = 'user@gmail.com';
+    delete process.env.EMAIL_FROM;
+    const sendMail = jest.fn().mockResolvedValue();
+    nodemailer.createTransport.mockReturnValue({ sendMail });
+
+    const mailer = require('../src/utils/mailer');
+
+    await mailer.sendMail({ to: 'a@test.com', subject: 'subject', text: 'body', html: '<p>body</p>' });
+
+    expect(sendMail).toHaveBeenCalledWith({
+      from: process.env.EMAIL_USER,
+      to: 'a@test.com',
+      subject: 'subject',
+      text: 'body',
+      html: '<p>body</p>',
+    });
+  });
+
+  test('sendMail uses EMAIL_FROM when defined', async () => {
+    process.env.EMAIL_USER = 'user@gmail.com';
+    process.env.EMAIL_FROM = 'Farm <farm@example.com>';
+    const sendMail = jest.fn().mockResolvedValue();
+    nodemailer.createTransport.mockReturnValue({ sendMail });
+
+    const mailer = require('../src/utils/mailer');
+
+    await mailer.sendMail({ to: 'a@test.com', subject: 'subject', text: 'body', html: '<p>body</p>' });
+
+    expect(sendMail).toHaveBeenCalledWith({
+      from: process.env.EMAIL_FROM,
+      to: 'a@test.com',
+      subject: 'subject',
+      text: 'body',
+      html: '<p>body</p>',
+    });
+  });
+});
